refactor: separate module protocol logic
